Correct gen_message input type Distro A, OPSEC #2893

+++ src/rust_pubsub/src/rust_helper/src/lib.rs
@@ -1,10 +1,10 @@
 #[cxx::bridge]
 mod ffi {
     extern "Rust" {
-        fn gen_message(count: usize) -> String;
+        fn gen_message(count: u32) -> String;
     }
 }
 
-fn gen_message(count: usize) -> String {
+fn gen_message(count: u32) -> String {
     format!("Hello from rust! {}", count)
 }
